refactor; moved dict classes
