+++ cognite/neat/_rules/models/dms/_schema.py
@@ -25,6 +25,13 @@ from cognite.client.data_classes.data_modeling.views import (
 )
 from cognite.client.data_classes.transformations.common import Edges, EdgeType, Nodes, ViewInfo
 
+from cognite.neat._client.data_classes.data_modeling import (
+    CogniteResourceDict,
+    ContainerApplyDict,
+    NodeApplyDict,
+    SpaceApplyDict,
+    ViewApplyDict,
+)
 from cognite.neat._issues import NeatError
 from cognite.neat._issues.errors import (
     NeatYamlError,
@@ -42,13 +49,8 @@ from cognite.neat._issues.warnings import (
 from cognite.neat._issues.warnings.user_modeling import DirectRelationMissingSourceWarning
 from cognite.neat._rules.models.data_types import _DATA_TYPE_BY_DMS_TYPE
 from cognite.neat._utils.cdf.data_classes import (
-    CogniteResourceDict,
-    ContainerApplyDict,
-    NodeApplyDict,
     RawTableWrite,
     RawTableWriteList,
-    SpaceApplyDict,
-    ViewApplyDict,
 )
 from cognite.neat._utils.cdf.loaders import ViewLoader
 from cognite.neat._utils.rdf_ import get_inheritance_path
