+++ tests/tests_unit/rules/test_importers/test_dms_importer.py
@@ -4,11 +4,11 @@ from typing import cast
 import pytest
 from cognite.client import data_modeling as dm
 
+from cognite.neat._client.data_classes.data_modeling import ContainerApplyDict, SpaceApplyDict, ViewApplyDict
 from cognite.neat._rules.exporters import DMSExporter
 from cognite.neat._rules.importers import DMSImporter, ExcelImporter
 from cognite.neat._rules.models import DMSRules, DMSSchema, RoleTypes
 from cognite.neat._rules.transformers import DMSToInformation, ImporterPipeline, VerifyDMSRules
-from cognite.neat._utils.cdf.data_classes import ContainerApplyDict, SpaceApplyDict, ViewApplyDict
 from tests.config import DOC_RULES
 from tests.data import windturbine
