+++ tests/tests_unit/rules/test_models/test_dms_schema.py
@@ -7,6 +7,7 @@ from _pytest.mark import ParameterSet
 from cognite.client import data_modeling as dm
 from cognite.client.data_classes import DatabaseWrite, DatabaseWriteList, TransformationWrite, TransformationWriteList
 
+from cognite.neat._client.data_classes.data_modeling import ContainerApplyDict, SpaceApplyDict, ViewApplyDict
 from cognite.neat._issues import NeatError, NeatIssue, NeatWarning
 from cognite.neat._issues.errors import PropertyNotFoundError, ResourceDuplicatedError, ResourceNotFoundError
 from cognite.neat._issues.warnings import FileTypeUnexpectedWarning
@@ -14,11 +15,8 @@ from cognite.neat._issues.warnings.user_modeling import DirectRelationMissingSou
 from cognite.neat._rules.models import DMSSchema
 from cognite.neat._rules.models.dms import PipelineSchema
 from cognite.neat._utils.cdf.data_classes import (
-    ContainerApplyDict,
     RawTableWrite,
     RawTableWriteList,
-    SpaceApplyDict,
-    ViewApplyDict,
 )
