+++ tests/tests_unit/test_utils/test_cdf_classes.py
@@ -1,6 +1,6 @@
 from cognite.client.data_classes.data_modeling import NodeApply, NodeId
 
-from cognite.neat._utils.cdf.data_classes import NodeApplyDict
+from cognite.neat._client.data_classes.data_modeling import NodeApplyDict
 
 
 class TestCogniteResourceDict:
